Singleton Design pattern implementation
Singleton class is created so that all of the file system contains a
single instance and that all other classes knows about the state of the
file system. It maintains the number of inodes allocated, the file name
and corresponding inodes for the entry 

It holds the quit command implementation. When q command is pressed the
root node begins to be written to the file system. 
Based on the inode count the files are written.